% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\name{sliding_hist}
\alias{sliding_hist}
\title{Sliding Histogram}
\usage{
sliding_hist(vec, bin_width, range)
}
\arguments{
\item{vec}{vector of numeric values.}

\item{bin_width}{width of bin to slide, same as histogram bin width}

\item{range}{where to start and stop sliding of bin.}
}
\description{
Takes one dimensional data and returns a data frame in the form of x value and sum which are the x and y coordinates of the sliding histogram respectively.
}
\examples{
sliding_hist(rnorm(100), bin_width = 0.1)
}

