% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\name{timelinePlot}
\alias{timelinePlot}
\title{Timeline Plot}
\usage{
timelinePlot(d, event_vars, event_labels = NULL, time_interval = "days",
  max_time = NULL, interval_width = NULL, greyscale = F,
  custom_xbreaks = NULL, custom_xlabs = NULL)
}
\arguments{
\item{d}{data for events}

\item{event_vars}{Vector of the titles of the columns containing event data}

\item{event_labels}{Vector of the labels for the event data. Must be in same order as event_vars. Defaults to event_vars.}

\item{time_interval}{Interval unit you want for the x-axis. Options are "days", "weeks", "years". If an unknown value is seen it defaults to days.}

\item{max_time}{How far in time do you want to the chart drawn. If not specified defaults to maximum seen event time.}

\item{interval_width}{Size of rolling interval for sliding histogram}

\item{greyscale}{Turn plot output into greyscale.}

\item{custom_xbreaks}{Vector containing breakpoints for the x-axis. If you want the have custom breaks on x axis to call out individual times.}

\item{custom_xlabs}{labels for custom x breaks, must match custom_xbreaks in length.}
}
\description{
Takes in wide format
}
\examples{
timelinePlot(my_time_data)
}

