% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\name{timelinePlot_slide}
\alias{timelinePlot_slide}
\title{Timeline Plot}
\usage{
timelinePlot_slide(d, event_vars, event_labels = NA, time_interval = "days",
  max_time = NA, interval_width = NA, greyscale = F, custom_xbreaks = F,
  custom_xlabs = F)
}
\arguments{
\item{d}{data for events}

\item{event_vars}{title of columns containing event data}

\item{event_labels}{Legend labels for the events}

\item{time_interval}{intervals you want for the x-axis. Options are "days", "weeks", "years". If an unknown value is seen it defaults to days.}

\item{max_time}{maximum time plotted}

\item{interval_width}{Size of rolling interval for sliding histogram}

\item{greyscale}{Want to output in greyscale?}

\item{custom_xbreaks}{If you want the have custom breaks on x axis to call out individual times}

\item{custom_xlabs}{labels for custom x breaks, must match in length.}
}
\description{
Takes in wide format
}
\examples{
timelinePlot(my_time_data)
}

