% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plot.R
\name{timelinePlot}
\alias{timelinePlot}
\title{Timeline Plot}
\usage{
timelinePlot(time_data, x_limits = NA, variable_name = "variable",
  interval_val_name = "value", time_interval = "days", title = "",
  alpha = 0.3)
}
\arguments{
\item{time_data}{dataframe with at least two columns, what's being measured and the time of that measurement.}

\item{x_limits}{Limits of the x-axis. Defaults to ggplot's scaling but if you want to zoom in or out do it here.}

\item{variable_name}{The name of the column containing the  name of the measurement}

\item{interval_val_name}{The name of the column containing the time to measurement}

\item{time_interval}{intervals you want for the x-axis. Options are "days", "weeks", "years". If an unknown value is seen it defaults to days.}

\item{title}{What you want displayed in the title section for the plot.}

\item{alpha}{Opacity level of the curves.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Takes data in with at least two columns- the measurement variable- and the time from baseline to that value.
}

